# docs/evidencias/05-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RepMatch.Catalogo.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5081/"
    }
  ],
  "paths": {
    "/salud": {
      "get": {
        "tags": [
          "RepMatch.Catalogo.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/busquedas": {
      "get": {
        "tags": [
          "Busquedas"
        ],
        "parameters": [
          {
            "name": "cantidad",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusquedaDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Busquedas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearBusquedaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearBusquedaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearBusquedaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusquedaDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/busquedas/{id}": {
      "get": {
        "tags": [
          "Busquedas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusquedaDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/busquedas/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Busquedas"
        ],
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusquedaDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearClienteDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearClienteDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearClienteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/clientes/{id}/vehiculos": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/repuestos": {
      "get": {
        "tags": [
          "Repuestos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepuestoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/repuestos/compatibles": {
      "get": {
        "tags": [
          "Repuestos"
        ],
        "parameters": [
          {
            "name": "marca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modelo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anio",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "motor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sistema",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepuestoDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/repuestos/{codigo}": {
      "get": {
        "tags": [
          "Repuestos"
        ],
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepuestoDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AplicacionDto": {
        "required": [
          "marca",
          "modelo",
          "anioDesde",
          "anioHasta"
        ],
        "type": "object",
        "properties": {
          "marca": {
            "type": "string"
          },
          "modelo": {
            "type": "string"
          },
          "anioDesde": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "anioHasta": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "motor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BusquedaDto": {
        "required": [
          "id",
          "clienteId",
          "vehiculo",
          "textoLibre",
          "estado",
          "fechaCreacion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "vehiculo": {
            "$ref": "#/components/schemas/VehiculoDto"
          },
          "textoLibre": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "codigosObjetivo": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ofertas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaDto"
            }
          }
        }
      },
      "ClienteDto": {
        "required": [
          "id",
          "nombre",
          "email",
          "fechaAlta"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nombre": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "fechaAlta": {
            "type": "string",
            "format": "date-time"
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehiculoClienteDto"
            }
          }
        }
      },
      "CrearBusquedaDto": {
        "required": [
          "clienteId",
          "vehiculo",
          "textoLibre"
        ],
        "type": "object",
        "properties": {
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "vehiculo": {
            "$ref": "#/components/schemas/VehiculoDto"
          },
          "textoLibre": {
            "maxLength": 1000,
            "minLength": 5,
            "type": "string"
          }
        }
      },
      "CrearClienteDto": {
        "required": [
          "nombre",
          "email"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "email": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "OfertaDto": {
        "required": [
          "codigoRepuesto",
          "nombreTienda",
          "titulo",
          "precio",
          "moneda",
          "precioTotal",
          "urlOriginal",
          "disponible",
          "capturadaEn"
        ],
        "type": "object",
        "properties": {
          "codigoRepuesto": {
            "type": "string"
          },
          "nombreTienda": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "precio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "moneda": {
            "type": "string"
          },
          "costoEnvio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "precioTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "urlOriginal": {
            "type": "string"
          },
          "disponible": {
            "type": "boolean"
          },
          "capturadaEn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RepuestoDto": {
        "required": [
          "id",
          "codigoCanonico",
          "nombre",
          "sistema"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigoCanonico": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "sistema": {
            "type": "string"
          },
          "codigosEquivalentes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aplicaciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AplicacionDto"
            }
          }
        }
      },
      "VehiculoClienteDto": {
        "required": [
          "id",
          "alias",
          "datos"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "alias": {
            "type": "string"
          },
          "datos": {
            "$ref": "#/components/schemas/VehiculoDto"
          },
          "vin": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "VehiculoDto": {
        "required": [
          "marca",
          "modelo",
          "anio"
        ],
        "type": "object",
        "properties": {
          "marca": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "modelo": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "anio": {
            "maximum": 2100,
            "minimum": 1950,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "motor": {
            "maxLength": 50,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "RepMatch.Catalogo.Api"
    },
    {
      "name": "Busquedas"
    },
    {
      "name": "Clientes"
    },
    {
      "name": "Repuestos"
    }
  ]
}